requirements.txt
openai-whisper
cohere
streamlit